feat: add AdminSchoolClassController and enhance user profile updates
- **Backend**:
  - Introduced `AdminSchoolClassController` with CRUD operations and user enrollment management for school classes.
  - Enhanced `UserService` to support controlled profile updates, with validation for username/email uniqueness and password verification.
  - Improved error handling in role-to-class-role mapping and refactored school class repository checks for better performance.

- **Misc**:
  - Simplified pagination logic and default size handling in `SchoolClassController`.
  - Updated DTOs (`UserRequest`) to include `currentPassword` for password changes during profile updates.
