+++ src/main/java/org/example/projectbackendteammycodebasebringsalltheboys/dto/user/UserRequest.java
@@ -12,6 +12,8 @@ public class UserRequest {
 
   private String password; // Optional for updates, required for creation
 
+  private String currentPassword; // Required for password changes in profile update
+
   @NotBlank private String roleName; // Name of the role, e.g., "ROLE_ADMIN", "ROLE_TEACHER"
 
   private java.util.List<java.util.UUID> schoolClassIds;
